Book Nook and Add Book working...
When adding a book, it must already exist within the pl_book table of the database. The case where the book does not yet exist in the database will have to interface with the ISBNdb.com API and fill in the remaining fields.

Increased size of search bar on style.css and decreased padding; it was covering up text on my browser.

Switched .html pages to .php and included session_start where necessary.