Pages that require login now redirect...
The user to the log_in.php page if they are not logged in. This prevents users from manually typing in an address and exploiting vulnerabilities with the website. And it just looks more professional than all of the error messages that would pop up without vars being initialized through the login process.